+++ packages/expo-maps/CHANGELOG.md
@@ -7,6 +7,7 @@
 ### 🎉 New features
 
 - Add `Circle` type to the `GoogleMaps` namespace. ([#49124](https://github.com/expo/expo/pull/49124) by [@CatLover01](https://github.com/CatLover01))
+- Add `tilt` and `bearing` to map camera positions.
 
 ### 🐛 Bug fixes